Fixed a delete sync bug.
git-svn-id: svn://svn.code.sf.net/p/ifolder/code/branches/WebSync@2383 db2431b9-df06-0410-971c-a38ea9a75485

+++ simias/src/Sync/FileDelta.cs
@@ -137,7 +137,8 @@ namespace Simias.Sync
 				workStream = null;
 			}
 			// We need to delete the temp file.
-			File.Delete(workFile);
+			if (workFile != null)
+				File.Delete(workFile);
 		}
 
 		#endregion
